requirements/requirements.txt
aniso8601==1.3.0
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
py==1.4.34
pytest==3.2.2
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
Werkzeug==0.12.2
Flask-SQLAlchemy==2.2
Flask-Migrate==2.1.1
marshmallow==2.13.6
marshmallow-sqlalchemy==0.13.1
flask-marshmallow==0.8.0
Flask-Script==2.0.6
python-decouple==3.1
